Refactoring e risoluzione issue segnalate da sonar
