+++ app/src/main/java/it/unimib/brewday/ui/AttrezziFragment.java
@@ -1,4 +1,4 @@
-package it.unimib.brewday;
+package it.unimib.brewday.ui;
 
 import android.os.Bundle;
 
@@ -12,11 +12,6 @@ import android.view.ViewGroup;
 
 import it.unimib.brewday.R;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link AttrezziFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
 public class AttrezziFragment extends Fragment {
 
 
@@ -32,22 +27,11 @@ public class AttrezziFragment extends Fragment {
         return new AttrezziFragment();
     }
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_attrezzi, container, false);
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
     }
 }
