+++ app/src/main/java/it/unimib/brewday/ui/ListaRicetteFragment.java
@@ -1,4 +1,4 @@
-package it.unimib.brewday;
+package it.unimib.brewday.ui;
 
 import android.os.Bundle;
 
@@ -10,11 +10,6 @@ import android.view.ViewGroup;
 
 import it.unimib.brewday.R;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link ListaRicetteFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
 public class ListaRicetteFragment extends Fragment {
 
 
@@ -25,11 +20,6 @@ public class ListaRicetteFragment extends Fragment {
     public static ListaRicetteFragment newInstance() {
 
         return new ListaRicetteFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
     }
 
     @Override
