+++ app/src/main/res/layout/fragment_crea_ricetta.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="it.unimib.brewday.CreaRicettaFragment">
+    tools:context="it.unimib.brewday.ui.CreaRicettaFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
