+++ app/src/main/res/layout/fragment_ingredienti.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="it.unimib.brewday.IngredientiFragment">
+    tools:context="it.unimib.brewday.ui.IngredientiFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
