+++ app/src/main/res/layout/fragment_lista_ricette.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="it.unimib.brewday.ListaRicetteFragment">
+    tools:context="it.unimib.brewday.ui.ListaRicetteFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
