+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,27 +7,27 @@
 
     <fragment
         android:id="@+id/listaRicetteFragment"
-        android:name="it.unimib.brewday.ListaRicetteFragment"
+        android:name="it.unimib.brewday.ui.ListaRicetteFragment"
         android:label="fragment_lista_ricette"
         tools:layout="@layout/fragment_lista_ricette" />
     <fragment
         android:id="@+id/birreFragment"
-        android:name="it.unimib.brewday.BirreFragment"
+        android:name="it.unimib.brewday.ui.BirreFragment"
         android:label="fragment_birre"
         tools:layout="@layout/fragment_birre" />
     <fragment
         android:id="@+id/creaRicettaFragment"
-        android:name="it.unimib.brewday.CreaRicettaFragment"
+        android:name="it.unimib.brewday.ui.CreaRicettaFragment"
         android:label="fragment_crea_ricetta"
         tools:layout="@layout/fragment_crea_ricetta" />
     <fragment
         android:id="@+id/attrezziFragment"
-        android:name="it.unimib.brewday.AttrezziFragment"
+        android:name="it.unimib.brewday.ui.AttrezziFragment"
         android:label="fragment_attrezzi"
         tools:layout="@layout/fragment_attrezzi" />
     <fragment
         android:id="@+id/ingredientiFragment"
-        android:name="it.unimib.brewday.IngredientiFragment"
+        android:name="it.unimib.brewday.ui.IngredientiFragment"
         android:label="fragment_ingredienti"
         tools:layout="@layout/fragment_ingredienti" />
 </navigation>
